backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# MT5 integration
MetaTrader5==5.0.5120

# Redis for real-time communication
redis[hiredis]==5.0.1

# WebSocket support
websockets==12.0

# CORS middleware
python-multipart==0.0.6

# Enhanced security and monitoring
psutil==5.9.6
cryptography==41.0.7

# Performance monitoring (optional)
prometheus-client==0.19.0
